feat(#302): Add new terms: exists, absent, size, and type to Factbase
