+++ test/factbase/terms/test_absent.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+# SPDX-FileCopyrightText: Copyright (c) 2024-2025 Yegor Bugayenko
+# SPDX-License-Identifier: MIT
+
+require_relative '../../test__helper'
+require_relative '../../../lib/factbase/term'
+require_relative '../../../lib/factbase/terms/absent'
+
+# Test for 'absent' term.
+# Author:: Yegor Bugayenko (yegor256@gmail.com)
+# Copyright:: Copyright (c) 2024-2025 Yegor Bugayenko
+# License:: MIT
+class TestAbsent < Factbase::Test
+  def test_absent
+    t = Factbase::Absent.new([:foo])
+    refute(t.evaluate(fact('foo' => 41), [], Factbase.new))
+    assert(t.evaluate(fact('bar' => 41), [], Factbase.new))
+  end
+end
